Rename assertType.mjs to inspectType.mjs
