+++ src-runtime/index.mjs
@@ -1,9 +1,9 @@
 export * from './assertMode.js';
-export * from './assertType.mjs';
 export * from './createDiv.mjs';
 export * from './customTypes.mjs';
 export * from './customValidations.mjs';
 export * from './div.mjs';
+export * from './inspectType.mjs';
 export * from './makeJSDoc.mjs';
 export * from './options.mjs';
 export * from './partition.js';
